+++ src/only-token-exchange.js
@@ -1,0 +1,7 @@
+import { NativeModules } from 'react-native';
+
+const { RNAppAuth } = NativeModules;
+
+export default () => {
+  return RNAppAuth.onlyTokenExchange();
+};
