+++ src/validators.js
@@ -1,0 +1,24 @@
+import invariant from 'invariant';
+
+export const validateIssuerOrServiceConfigurationEndpoints = (issuer, serviceConfiguration) =>
+  invariant(
+    typeof issuer === 'string' ||
+      (serviceConfiguration &&
+        typeof serviceConfiguration.authorizationEndpoint === 'string' &&
+        typeof serviceConfiguration.tokenEndpoint === 'string'),
+    'Config error: you must provide either an issuer or a service endpoints'
+  );
+export const validateIssuerOrServiceConfigurationRevocationEndpoint = (
+  issuer,
+  serviceConfiguration
+) =>
+  invariant(
+    typeof issuer === 'string' ||
+      (serviceConfiguration && typeof serviceConfiguration.revocationEndpoint === 'string'),
+    'Config error: you must provide either an issuer or a revocation endpoint'
+  );
+export const validateClientId = clientId =>
+  invariant(typeof clientId === 'string', 'Config error: clientId must be a string');
+
+export const validateRedirectUrl = redirectUrl =>
+  invariant(typeof redirectUrl === 'string', 'Config error: redirectUrl must be a string');
